backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.11.0
httpx==0.28.1
sentence-transformers==5.1.0
torch==2.8.0
rq==2.4.1
redis==6.4.0
SQLAlchemy==2.0.43
asyncpg==0.30.0
aiosqlite==0.21.0
psycopg2==2.9.10
python-dotenv==1.1.1
langchain==0.3.27
langgraph==0.6.6
unidiff==0.7.5
alembic==1.16.5
bcrypt==4.3.0
cryptography==45.0.7
pgvector==0.3.6
